refactor: move field to path mapping to validation layer
