+++ src/test/kotlin/se/uulm/snowballr/backend/arch/LayerArchitectureTest.kt
@@ -201,6 +201,7 @@ private class NamingConventions {
             .orShould(haveSimpleNameEndingWith("ValidatorKt")) // kotlin class
             .orShould(haveSimpleName("ValidationHelperKt")) // exception
             .orShould(haveNameMatching($$".*inlined\\$zipOrAccumulate\\$.*")) // inlined function exception
+            .orShould(haveNameMatching($$".*\\$WhenMappings")) // synthetic class for `when` over an enum
             .orShould(haveNameMatching(".*validate.*")) // inlined function exception
             .because("All validators should have the 'Validator' suffix")
             .check(classes)
